[roll] Roll fuchsia [inspect] Wire up the dispatcher for inspect::MakeTreeHandler
Currently, inspect::MakeTreeHandler takes in an async_dispatcher_t
which it does not pass on to the BindingSet::AddBinding call.

Clients of the library need to set an async dispatcher to get around
this issue.

This fix correctly forwards the async dispatcher into the AddBinding
call, removing the need to set a dispatcher separately.

Test: `fx test inspect_service_cpp_tests` passed
      `fx shell runtests -n inspect` passed
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401497
Original-Revision: b9f2105ad8e7622df9dca537dbae0b74e87fe291
GitOrigin-RevId: 2d438a423b285c2d3752a52016b91417ace33481
Change-Id: I0f7ecd5fd42d7b6f1335ded8ad5e9ccbee38c7e1