[roll] Roll fuchsia [bt][sm] Clarify the state machine in the ScStage1 classes
Firstly, this CL cleans up a possible externally-triggered assertion in
the ScStage1JustWorksNumericComparison class from receiving multiple
Pairing Random values. The first Pairing Random message would, through
a series of callbacks, trigger the destruction of the ScStage1JustWorks
NumericComparison class by its owning class, so in practice, this would
not be exploitable. However, this depends on logic external to the
class, which is not desirable, so a check for this case has been added.
Secondly (and relatedly),this CL adds documentation and improves naming
to make the state flow of the ScStage1 classes more clear.

Original-Bug: 53589
Test: Added SMP_ScStage1JustWorksNumericComparisonTest.ResponderReceive
      RandomTwiceFails to `$ fx test bt-host-unittests`
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397367
Original-Revision: 5349b73a4a696627446526be3ab73ab2542c9922
GitOrigin-RevId: 7d60e064437730eff05833da2275fdbb51b000ef
Change-Id: Ib56850e7ed93178024fc0f89d6c00f62ae902bf7