[roll] Roll fuchsia [kernel] Zero the symbolizer file VMO if not publishing it
All of the instrumentation data VMOs are always published to
userboot for simplicity, but zero-length VMOs will be ignored in
userland.  The symbolizer file VMO was left live with a size of
PAGE_SIZE when it wasn't meant to be published, so userland saw
it as live but its name hadn't been set.  Instead, just reset it
to zero size so it will be ignored like all the others.

Original-Fixed: 54904
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401607
Original-Revision: ff5c0e9f1630afb6dc7877a10e20b76ef5c43075
GitOrigin-RevId: 53d59e5fd3b563361e8f82a32803710eaec9fceb
Change-Id: I4acbeb5f449cd2f417a413845c72c00208347db2