[roll] Roll fuchsia [smoke-test] Fix rustc_test's test spec in tests.json
Use the target_name in the context of rustc_test, rather than using
invoker.target_name inside the test_spec instantiation. This corrects
the tests.json label for the test to be the expected label to build the
gn test.

This changes result in diffs like the following in tests.json:

    <       "label": "//src/developer/development-bridge:ffx(//build/toolchain:host_x64)",
    ---
    >       "label": "//src/developer/development-bridge:ffx_bin_test(//build/toolchain:host_x64)",

for

    "test": {
      "command": [
        "host_x64/ffx_bin_test"
      ],
      "cpu": "x64",
      "label": "//src/developer/development-bridge:ffx_bin_test(//build/toolchain:host_x64)",
      "name": "host_x64/ffx_bin_test",
      "os": "linux",
      "path": "host_x64/ffx_bin_test",
      "runtime_deps": "host_x64/gen/src/developer/development-bridge/ffx_bin_test_spec.deps.json"
    }

Confusingly, while the ':ffx' target does exist, it's not a test:

    [rust-tests scottmg@around /work/fuchsia]$ fx gn outputs out/default '//src/developer/development-bridge:ffx(//build/toolchain:host_x64)'
    host_x64/obj/src/developer/development-bridge/ffx.stamp
    [rust-tests scottmg@around /work/fuchsia]$ fx gn outputs out/default '//src/developer/development-bridge:ffx_bin_test(//build/toolchain:host_x64)'
    host_x64/ffx_bin_test

which confuses `gn analyze` when it tries to output which tests are
affected by a source change. So I think this was the original intention
of the rustc_test template.

Original-Bug: 53604
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401622
Original-Revision: 105ae9a8b256db654613f42c531167742cdbc359
GitOrigin-RevId: 1d37fa36d4abbdaab0de15721ed23e12a4d285d8
Change-Id: I84e3e0272a1ae66cb1c239dcb61df604f5012222