[roll] Roll fuchsia [security] remove appmgr and fshost from main_process_critical
Both fshost and appmgr were added to the main_process_critical
allowlist, but neither request the feature yet because there is upcoming
code work to get them to cleanly exit when asked to stop. Remove them
for now, and we'll readd them once the components are ready to take
advantage of the feature.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401672
Original-Revision: 50404baf1376991b0be6d31ab592dd95fd5c8589
GitOrigin-RevId: d1937eeb43e7ba26ea7d5b7446a1e5afa13e86a1
Change-Id: I6192f18f8e76b935081c2bbc0a4c0a5de390ff1c