[roll] Roll fuchsia [wlan] Remove the definition of WlanInfoMacRoleMask
Adding WlanInfoMacRoleMask was a mistake. The enum type
generated by Banjo is intentionally overloaded for bitmasks in Rust
and C++. In general, adding *Mask types will not work in Rust code
because Rust sees the *Mask type as different from its sister * type.
The semantic benefit of adding the *Mask is very small and is
therefore removed.

Note that in the case of WlanInfoMacRoleMask, there are no
issues compiling. There are however issues with creating a type such
as WlanInfoHardwareCapabilityMask because Rust performs type checking
on case statements using the type and does not find any values for it.

Original-Bug: 52941
Test: No additional tests necessary
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400811
Original-Revision: 18a3444c65ea5fe08d8880430c08b5f688d39f63
GitOrigin-RevId: 63193c0cec7c5588db833c49e0dcb3623a04d202
Change-Id: Iad5c4190d68e4e3f9545c60c89dc9e5b0b091000