[roll] Roll fuchsia [shutdown-shim] reduce statecontrol.Admin timeout
The shutdown-shim's timeout for power_manager is always waited for on
bringup, as power_manager will never finish loading on bringup. To allow
developers to reboot in < 20 seconds on bringup, reduce this timeout to
2 seconds. In all typical shutdown scenarios power_manager will already
be running, so it should take far less than 2 seconds for the connection
to be established with power_manager.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400815
Original-Revision: 62b4a993c5112de8bbc988c25490e7df5de8c187
GitOrigin-RevId: 549427b8bdbc350ddbf0c452f47686a3f977d0da
Change-Id: Idf0d59a1210e6f85ade5c4aa5eb7dbacdf4f4902