[roll] Roll fuchsia [fidl][benchmarks] Remove roundtrip benchmarks
Roundtrip benchmarks can be useful in general, but the current
benchmarks send messages in one binding (Rust) and then echo in another
so it is hard to understand performance changes. As a result we haven't
actually been using these for tracking regressions, but they add noise
in the results.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401116
Original-Revision: 75039f69a3ccb5247e17253e616f9bd5c1be1560
GitOrigin-RevId: d4f38a9a20b9dbae85318f17464f8b5d2183f6e3
Change-Id: Ic17752505c32e287c4d0eac439abe1e3b264fc1a