[roll] Roll fuchsia [location] introduce BssCache trait and RealBssCache struct
Refactor the BSS cache, by moving the public interface of the
cache into a `trait`, and renaming the implementing `struct`
to `RealBssCache`.

This refactoring will enable the main-loop to use a test
double in place of a `RealBssCache`. This will:
1. Simplify the main loop tests, by abstracting away
   interactions with the `ScanResultIterator` FIDL service.
2. Isolate main-loop tests from any logic errors in the
   `RealBssCache`. (Any logic errors in `RealBssCache`
   should be found by _its_ unit-tests, not the tests
   for main-loop.)

Original-Bug: None
Test: fx set core.astro --release --with //src/connectivity/location/emergency \
      && fx format-code \
      && fx build \
      && fx test emergency-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393322
Original-Revision: ba886edfed65708aec2163e3fd74c6e80f6f733a
GitOrigin-RevId: 783b44a92e6d9547fa3a9a5e890716a98583a2a0
Change-Id: I92b2c0a91c7ba8bbbc5f52c9d16c3a473e7f79d5