[roll] Roll fuchsia [tee][optee] Lower log level of OpteeClient logs
The OpteeClient acts as a proxy between a Trusted App and a Client App.
If the TEE OS, Trusted App or Client App provide malformed or
unsupported messages, this does not indicate a bug in the driver. As
such, the driver should rarely be using the ERROR level for these logs.
This CL updates the log statements in OpteeClient to mostly use the
DEBUG level.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401976
Original-Revision: 987fe36fb47962d16422559ddce3b7baad8bb8d6
GitOrigin-RevId: 09213580cc4b70f3b80fb400f54335925788d927
Change-Id: I746d47142c31eeb9fc595f92e096859c4a9c1698