[roll] Roll fuchsia [fx][test] fix hash pinning update on build
fxtest was fetching the merkle root hash in a wrong location, before
the build was attempted. As a consequence, it was using the hash of the
package before the build, hence executing stale tests.

Implementation details of this fix: executionhandle is the objec that
contains the exact Fuchsia URL of the component test. However, it was
being created at testdefinition construction time, which is way before
the 'fx build' has a chance to run. In order to fix the issue, this CL:
- moves testType from ExecutionHandle to TestDefinition
- adds a "hash" public non-final attribute to TestDefinition
- moves ExecutionHandle from being an attribute set at construction time
in TestDefinition to being a public method that calculates it at
execution time, and makes that method use the "hash" that is set at the
moment of execution (via the new decoratedPackageUrl getter)

As a side effect, 'fxPath' is not needed anymore in TestDefinition.
Now, instead of having the full path of fx in CommandTokens, only the
string 'fx' is used, and TestBundle._runTest does two new transformations:
replace 'fx' with FxLocator.fxPath if 'fx' is the first element of the
CommandTokens, and query the PackageRepository for the
testBundle.testDefinition packageName and set it in testDefinition.hash.

Original-Fixed: 54881
Tests:
  - added a new integration test in fx self-test to verify this behavior
  - manually tested with a component test
  - 'fx test fxtest'
  - 'fx self-test subcommands/fx_test_test'

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401794
Original-Revision: 4a6a094395ac29e76a117d4b33f54ed71092b444
GitOrigin-RevId: e2e0d6a7563c88a306c0e63840c800ead2ced97a
Change-Id: I8b2f988d72460aec9ea50ca1ee13cbf113f7a512