[roll] Roll fuchsia [storage] Move minfs tests to new fs_test framework.
Also fixed file_system -> filesystem and a header path that was wrong.

Tweaked the test slightly so that it runs a little faster which will
allow us to always run the FullOperations test (which was previously
marked as large so didn't run in CQ), and the test actually got broken
recently (when the minimum inode count was reduced) and will no longer
run on the non-FVM version of the test. That probably doesn't matter; we
have enough coverage from the FVM version of the test.

Original-Bug: 53179
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401893
Original-Revision: e3609e1b93b4fb433ca7a271756f8da4731a0214
GitOrigin-RevId: 7a8f1c1e066b6c58810aecb2c81d9e5394cfa92f
Change-Id: Iad3c58fb84bf4adc488ccef5bad784b96a38a8da