[roll] Roll fuchsia [mdns] fix build when enable_mdns_trace = true
enable_mdns_trace compiles in the formatting code required for the
mdns service to log mdns traffic. This CL fixes a build break that
occurs when enable_mdns_trace is set to true. The problem was the
redefinition of an operator<< overload.

Test: no behavioral change
      build with enable_mdns_trace = true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402215
Original-Revision: 125136c712b8325d15f35fcb73c731c7d62815f0
GitOrigin-RevId: 3dbe2107ae838bda8960928e9e9fa87827be68de
Change-Id: Ifab33d3287a0ff2c123aa9854c7a7f037df2a91d