[roll] Roll fuchsia [fidl] Move non-nullable but absent logic to visitor
This removes kAllowNonNullableCollectionsToBeAbsent and replaces it with
a visitor method that is called when a non-nullable absent collection is
encountered.

This has a few effects:
- VisitPointer now only handles present pointers
- Separates out the special case of null string/vectors in LLCPP and
makes clearer what happens in this case.
- Opens the door for code size reductions in encoding.cc if ever needed,
because the linearize_and_encode and encode cases can be controlled by
run-time if statements. That said, this probably isn't advisable in
practice because it is likely that we get performance benefits by not
needing to check whether linearization should be performed at each step.

Original-Bug: 53258
Test: Uses existing tests, e.g. fx test fidl-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401913
Original-Revision: 46539743f845eb45721f1539e46436e95127b89a
GitOrigin-RevId: 86b9a65b3aaabe25828462731dcbc8515513db0e
Change-Id: If7ffe186801e94f4958bf2fbcb4ffef631953e43