[roll] Roll fuchsia [tee_manager] Remove TeeDeviceConnection
The TeeDeviceConnection struct wrapped a DeviceConnectorProxy with an
AbortContext to enable registrants to abort their futures if the
channel to the DeviceConnectorProxy was closed. There was one
abort registrant for the ServiceFs and an additional abort registrant
for each passthrough connection to the tee.Device served by the device
driver. This CL restructures the tee_manager to not use
TeeDeviceConnection or AbortContexts to instead just do a select! on the
services future and the DeviceConnectorEventStream future so that the
tee_manager can shutdown appropriately.

Original-Fixed: 51733

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399154
Original-Revision: dbd9d48d6f9f02cc13aad3c0182701d60a2246dd
GitOrigin-RevId: f7f7e3d601f0cae1f19dd2c525547afd2082f329
Change-Id: I43fbbf366f2be563f27c7a560b8c711d0a87993b