[roll] Roll fuchsia [dart] allow dart_tool to come from a snapshot
We will be moving to use the prebuilt snapshot for the
dart kernel compiler instead of the custom one in the
Fuchsia tree. This target will allow for a dart_tool to
be built from a snapshot instead of sources.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402565
Original-Revision: 3f61fb42b460c1cc213758bf4bca1674a2132bb1
GitOrigin-RevId: a19e462440567c79fc56eb824a487ecbf905eb52
Change-Id: I465bd411b65b330f92dbdf5dcf9dad7770592551