[roll] Roll fuchsia [camera] Add ISP Global registers to ISP driver resources
The entire ISP MMIO region is mapped as cached
for performance optimization. This CL passes the
ISP global registers separately so that they can
be mapped as uncached to further optimize performance
by reducing the number of cache invalidations.

This CL only passes this MMIO region to the ISP
driver via the board driver, it is not used currently.

Test: fx test camera_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402505
Original-Revision: 51754ab6ce291e233ff15ed426a8cead40e67605
GitOrigin-RevId: 7efa3dbb40b8bb5fb8687052a4497bdb6a8f7258
Change-Id: Ic5657123d06675c43e5b90b5a93fda7bfc04559e