[roll] Roll topaz [flutter] use icudtl.dat file from futter prebuilt
The icudtl.dat file that the runner expects is different
from the one in fuchsia's repo. The file ends up in flutter sources
but we will eventually want to use one that is rolled in as a
prebuilt. We are taking a dependency on this file so we can
turn our hosts back on.

BUG: 54945

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/topaz/+/402173
Original-Revision: 33a5d14bc3c94d538c3953bd8b04bd107e2c4c95
GitOrigin-RevId: 60b3a75eab548c52f0f5c8fd2e60c5af8defa4ea
Change-Id: Ic1d495ebc15e59c1b2f056e5522eb31a34b36dbe