[roll] Roll fuchsia [blobfs] Wire up blobfs Inspect API
This CL wires up the blobfs Inspect API and
connects it to the diagnostics directory of fshost.

Archivist picks up blobfs and fshost diagnostics
and shows both with the `iquery show bootstrap/fshost`
command.

This is done by creating a RemoteDir subdirectory in the
fshost diagnostics directory and passing the server
end as an additional handle to the blobfs process.

blobfs::Runner serves the diagnostics directory
in addition to the root.

The additional handle passed in to blobfs is a
TEMPORARY measure. See fxb/54521 for more details

Test: 1. `iquery show bootstrap/fshost` now shows two
         trees, one for fshost and one for blobfs.
      2. Ran all tests -> `runtests /system/test/blobfs*`
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400753
Original-Revision: 09eeb3cf106171d97566c2cbc6733382a67eacd5
GitOrigin-RevId: 85278ba23c777a65cf59a78276c883055c5e8dcb
Change-Id: I2422d4d860659c43b996bf9f83755b20aa94512c