[roll] Roll fuchsia [check-system] Make src/developer/shell/mirror check-system clean
Originally, the error was that test_shared.h can't include
<lib/async-loop/default.h>. Because test_shared.h is included in
host-built code, it can't deps on async-loop-default. But it's not
actually needed in that header, so move it to the location where it's
needed (for &kAsyncLoopConfigNoAttachToCurrentThread), and the deps that
target on async-loop-default.

Original-Bug: 55166
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402919
Original-Revision: 7a3ef50c9e6bd3933629898bdca50989f9a8a90f
GitOrigin-RevId: a7d50108dabec252fb7d52df5c0d97f0018ca005
Change-Id: I67d24b13f71f36de176a19e132ae57072b728e90