[roll] Roll fuchsia [fidl] Fix binding_set capture handler move
`capture_handler` is captured by move for the move handler that is, by
default, const. Thus, the `std::move` of the `capture_handler` inside
the lambda is not actually moving, causing potential invalid copies of
the function.

This change fixes this issue by making the lambda mutable and moving to
the move-only `fit::function` to prevent this issue at compile-time in
the future.

Original-Bug: 55165
Original-Fixed: 8083

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402929
Original-Revision: 13e16e149fb2f4a9890e47de69c842df54530791
GitOrigin-RevId: c506357c3a85576110b86125f549ba9235dd97d6
Change-Id: I11fbd9434667795aba7421aebd90b2a14463308f