[roll] Roll fuchsia [mdns] make Mdns class unit-testable
This CL refactors the Mdns class a bit to make it more testable. The
Mdns class is essentially a complete mDNS implementation without the
FIDL service layer.

The key change to the Mdns class is to remove its embedded
MdnsTransceiver instance and allow the instance to be specified in
the Mdns constructor. This change includes introducing an abstract
class for transceivers (Mdns::Transceiver), which we can fake out in
tests.

A simple subscription unit test has been added as well as a
regression test for fxb/55116, currently disabled (because 55116
isn't fixed).

Original-Bug: 55116

Test: fx test mdns_tests
      Refactoring for unit test.
      Added new unit test for Mdns behavior for simple subscription.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402967
Original-Revision: 96592bb7797167d2548bea25703a55ca40afa69f
GitOrigin-RevId: 9fd36a602c9431db7733f3f268333c332860d528
Change-Id: I0a716f41f9312885fd885737fe111f12a6ebd026