[roll] Roll fuchsia [fidl] Fix lifetime of binding_set error handler
`BindingSet` sets the error handler of each `Binding` in the set to a
lambda that references the `Binding`. Previously, the removal and
destruction path of a `Binding` on error would:

(1) invoke the lambda,
(2) remove the `Binding` from the set,
(3) implicitly destruct the `Binding`'s error handler by setting it to
    nullptr,
(4) continue to reference some of the lambda's captured variables.

By doing step (3), the code has effectively destructed the lambda's
capture context. Then, step (4) continues to reference the lambda's
now invalid capture context. This is undefined behavior.

This change mitigates this issue by forcing the lifetime of the
`Binding`'s error handler to live until the very end of the lambda.
Additionally, to prevent future issues, the code now asserts that the
error handler is still set (and thus the running lambda is still set)
after the `RemoveOnError` code path is still run.

Original-Bug: 55165
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402971
Original-Revision: 8c2e0152fd8a73e6179a4c7e81d8981578811b63
GitOrigin-RevId: 88f2370019650f3b9a61b77bc69a9e018fe4a4e0
Change-Id: Idb05ebceba88f39bf8ae49c604dad90ce6df9ccd