[roll] Roll fuchsia [kernel][vmo] Fix decommit range checking for child slices.
Fix an issue that adanis@ spotted while looking at a different issue.
Essentially, when we perform a decommit range-op on a VMO which is a
child slice of another VMO, we need to be sure that we trim our range
against the child VMO, not it's root non-slice ancestor.

Test: new unit test
Original-Bug: 53547
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400157
Original-Revision: 3594eeb14b865cec057b5412e95d531d68e1939d
GitOrigin-RevId: 46523d7ceaf54c2fe592679e42543ed7636883a5
Change-Id: I2f0761dd54b1460b5bcd2068880e4508d496cadd