[roll] Roll external/github.com/catapult-project/catapult pinpoint: Break out details of runtime
This change breaks out the time a job was created, how long it's been in
the queue, when it's started, and how long it's been running (or how
long it ran for). We need to change the serialized model to include the
start time for each job, and handle the cases when a job is cancelled.

Unfortunately we don't have JavaScript testing set up for these elements
(yet) so we're going to have to see this in action in production and
follow up on any inconsistencies we find.

R=abennetts@google.com, fancl@chromium.org

Original-Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2266094
Original-Revision: 8353ff927b67c6484963d994532bff43ae65cd3c
GitOrigin-RevId: 393fe72be667e275ef3446b14a814389e5044a58
Change-Id: Ie04da220443fb030fb3e1c12c232c86a5e8313d3