[roll] Roll experiences [ermine] Add pumpUntilVisible utility function to app_widget_test
This change adds pumpUntilVisibile to WidgetTester using an extension.
This allows waiting for a widget to become visible. For now, it is
used only in one test, but would find a better place in a shared
library. That is left for subsequent cls based on need of this
method.

Test: All unittests pass
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/403198
Original-Revision: 0ca051497ae5f98989101038286c96de307b2c99
GitOrigin-RevId: dfe015c045bc329e8f2dd0d298997ffe7b9b15f0
Change-Id: I69a7eba78839586515408d85b1817216d786ef73