[roll] Roll fuchsia [blobfs] Remove redundant Verify() for paged blobs
When LoadAndVerifyBlob() is called on a paged blob, we simply need to
read in all the pages. An explicit call to Verify() is not required,
since the pager verifies all pages that are read in. The current code
verifies the blob twice.

Replace the call to Verify() with a zx_vmo_op_range(ZX_VMO_OP_COMMIT...)
on the blob's VMO, which will read in all the pages. Verification errors
encountered by the pager will be propagated back by the return status of
zx_vmo_op_range().

Test: Added BlobfsCheckerPagedTest to blobfs-unit.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402819
Original-Revision: a624e19cd6a77ba8302cc655cded76c54b77e859
GitOrigin-RevId: acf13f5365f81734f04e1465007cb132afd171a0
Change-Id: I57d909e85f3f5ca21d89d08c75493dbcc8a6425d