[roll] Roll fuchsia [a11y]  Initializes Speaker inside the Screen Reader Context.
This change initializes the Speaker inside the Screen Reader context.

In order to initialize the Speaker, we need the NodeDescriber, which in turn uses the MessageFormatter (effectively using the whole i18n stack built for the Screen Reader)

This is the first change in a series to migrate all actions to start using the Speaker.
The first step is to instantiate  the Speaker (this change), where a future change will move all actions to use it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402560
Original-Revision: c2db4c098af78b68d69f7f0af4d85c9022f7f0c5
GitOrigin-RevId: 2792787cda099c017354163eb254ec269e4fefe1
Change-Id: I3c95cf0a606fee801eff6059c6eda7a060c296c1