[roll] Roll fuchsia [storage] Add a wait method to ExecutionScope.
At the moment, it's not possible to implement a vfs server that can wait
for all connections to be closed, so this adds a wait message that can
be used.

Also changed vfs tests to run as a v2 test component rather than v1.

Original-Bug: 53796
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402722
Original-Revision: aab80d744fa163e46e111bbc8053af0e379a3ce9
GitOrigin-RevId: cf76841f8d8d1a4faa6cac57843b825fa13710d1
Change-Id: Ie1c5309a04efaca8b728650413c68545079240ad