[roll] Roll external/github.com/catapult-project/catapult sched: Support per-configuration job concurrency
This introduces an option for bot configurations to support Pinpoint job
scheduling parameters. We're setting this per-configuration to allow
Pinpoint administrators to associate costs to each kind of job and allow
cost-based packing of jobs depending on the dimensions of a pool.

The initial implementation uses budgets and costs for different
comparison modes. We're changing the scheduler to be aware of using
budgets and allowing multiple jobs to be scheduled concurrently to
consume the budget. We're designing for over-subscription which means we
can exhaust the budget with our priority-based scheduler.

The costs per job are assigned at creation time, because adding
retroactive cost-based scheduling will require we make operations on the
`bot_config` module to be transaction-aware and thus unnecessarily
increase the costs throughout its uses in Pinpoint. We'll do this for
now as a work-around until we can move the bot configurations to
git-hosted files watched by luci-config instead of being UI managed
configurations served via Datastore.

R=abennetts@google.com, fancl@chromium.org

Original-Bug: chromium:981429
Original-Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2272917
Auto-Submit: Dean Berris <dberris@chromium.org>
Original-Revision: 9e96ce18033151f6695012f97d7fd0de23278753
GitOrigin-RevId: 0ff4252e8d84abc2f2353d22b874f3384d02c459
Change-Id: Ieed11304fbfebc295837c20dc4716a3f66365592