[roll] Roll fuchsia [bt][emulator] Do not emulate advertising without LE support
A previous change (fxr/376913) made it possible for a fake LE peer to be
discoverable with empty advertising data. This caused a regression that
enabled advertisements to be reported for all fake peers, including
those that are designated to emulate BR/EDR and not LE.

This was because the presence of advertising and inquiry data was being
used to determine transport support. This is now changed to use the
address type instead.

Original-Bug: 55359
Test: This regression wasn't caught in CQ because bt-integration-tests
      is disable pending investigation into flakes. To test this change
      run:
          bt-integration-tests
          bt-host-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403756
Original-Revision: 9fe71dd2ac2fe3ff18f486d5189ad807aeb029f7
GitOrigin-RevId: d28ecf0c1afaacdc0f4fd6f2451ea09350b3f965
Change-Id: Ic2400f9aa4c4419d87082d1d77fc0c13273c85ca