[roll] Roll fuchsia [flatland] Remove pending operations.
Prior to this change, Flatland operations were typically not executed
until the client called Present(). The primary intention was that
changes to the scene should not be visible to users until the client
explicitly says they should be.

Some functions, such as those creating Links and registering buffer
collections with the Renderer, didn't even obey this rule because the
client might block on logic in those actions (receiving layout data,
being notified of sysmem buffer allocation, etc...)

The UberStruct now represents an atomic commit of the new state of
a particular Flatland instance. As a result, delaying operations
until Present() is no longer necessary.

Functions that delete Links are an exception to this rule, since the
destruction of a Link immediately results in changes in the
LinkSystem that could affect an about-to-be-rendered frame. As a
result, UnlinkFromParent() and ReleaseLink() still have some
functionality delayed until Present(). Specifically, Link objects
are not destructed until after a new UberStruct is published to
ensure that data in the LinkSystem can only ever be "older" than data
in UberStructs.

TEST=fx test -o flatland_unittests
MULTIPLY: flatland_unittests (fuchsia): 30

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398410
Original-Revision: 77bc77c79fe35924739521890e1162e62ab3ba11
GitOrigin-RevId: c096a553d7f59072cb9465d5e501f8f436724758
Change-Id: I98500a0765954ab9b4cc7996a564b50b5904d7d8