[roll] Roll fuchsia [kernel] Cleanup: Remove vmm_aspace_t in favor of directly using VmAspace
vmm_aspace_t was a C type that matched VmAspace (a C++-only type). It
was used as a field member in certain structures and as an API parameter
for (formerly) C code interacting with address spaces.

All of the headers and interfaces using vmm_aspace_t are now only
included or called from C++ code. Finish the conversion to VmAspace by
deleting the wrapper type.

Original-Bug: 55297 It would be nice if the Fuchsia build was faster
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403205
Original-Revision: 573594d75ce654e381a62f7c2c4e1bde2aabac82
GitOrigin-RevId: 8c8a8276c11189af963fda4d75dba73e7033beb9
Change-Id: Icf4e471720af655542b0ea4d1b1b3caad81591ac