[roll] Roll fuchsia [codec] Lower channel close logging to WARNING
Closing the client channel when vFail is called with is_fatal=false
is not an unrecoverable error. According to the log severity guidelines
the log for that case should not be an ERROR. Change to WARNING so
clients still knows why the channel closed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403760
Original-Revision: 55c4ef724cc8d55a75d437ae715cd24ea2494baa
GitOrigin-RevId: d08d838f5134b60d65d76fb3ba61981f59cce3be
Change-Id: Ibc9e5b362f19c4b330168c6b2f1e322f7e4e2900