[roll] Roll fuchsia [kernel] Cleanup: Enable -Wunused-functions in most sites it was disabled
Re-enable -Wunused-functions in:
* //zircon/kernel/arch/x86
* //zircon/kernel/kernel
* //zircon/kernel/platform/pc
* //zircon/kernel/vm:tests

This will help find unused code, which would show certain bugs (tests
not connected to the build) and help remove dead code.

Re-enabling this warning showed that there was a test in vm_unittest.cc,
vmpl_take_empty_test, which was not connected to the build. Add it to
unit tests.

After this there are two uses in-kernel -- one in vm, one in console.

Original-Bug: 55297 It would be nice if the Fuchsia build was faster
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403199
Original-Revision: 280249fc7f2a82b6e5d7f36221cd9e066355ec8c
GitOrigin-RevId: d51a362898c9cfcde8d1b94bf64e6ac11d5d9d37
Change-Id: I251b01364012e51287531646c9154f5d0fc6b126