[roll] Roll fuchsia [bt][test] Initial Profile Test Server impl
Creates a new component, bt-profile-test-server, that will be used to serve
the `fuchsia.bluetooth.bredr.ProfileTest` service. This component should
be used in a test setting only, and provides a way to receive updates about
a Bluetooth profile-under-test. The provided updates correspond to
expected profile behavior - such as ServiceFound notifications,
connection requests, and component termination.

The server enables a test client to launch a profile in a NestedEnvironment,
register and advertise services, search for services, and connect to
advertised services in the fake piconet.

A peer can be initialized with a PeerObserver relay, which will echo
any relevant events over the proxy. This is the main mechanism for
validating peer behavior.

Original-Bug: 51454
Test: Added bt-profile-test-server-tests

Multiply: bt-profile-test-server-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383846
Original-Revision: 40b950e4e26fa74cc188049bc02e1c322aa929f9
GitOrigin-RevId: e98d4cd6d5434855c352baac7cfb3a6575d46786
Change-Id: Idd672948469acd13218ab86384bd52bd66e1cc9a