[roll] Roll external/github.com/catapult-project/catapult exec: Use successive doubling for sample sizes
This change makes Pinpoint double the number of attempts up to a cap of
160 per change starting from 10 attempts when attempting to get stronger
statistical sampling. This means we get up to 4 iterations to get from
10 to 160 attempts with successive doubling.

We also change it so that if we decide to scale attempts for change[a]
that we'll also scale attempts for change[b] if they both have the same
number of attempts. This change should greatly reduce the amount of
waiting that's currently happening when we end up needing to get more
samples to get a stronger statistical signal on whether we should bisect
between two changes.

R=abennetts@google.com, fancl@chromium.org

Original-Bug: chromium:1101200
Original-Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2275571
Auto-Submit: Dean Berris <dberris@chromium.org>
Original-Revision: 96ad35b01419478d19cdfa17520efebc94148401
GitOrigin-RevId: 6335946bec45df6b69e4ee5d75ffbc112e3eabf5
Change-Id: I520d5ba7e31b962c274469aa6b9f67c80b74de4a