[roll] Roll fuchsia [system-update-checker] Clarify attempt ownership
Previously, the async task that owned an in-progress update attempt was
whichever task initiated that update attempt, which could either be the
initial check 1 minute after service start, a user initiated request via
FIDL, or the periodic background update check, which was accomplished
using shared state so the tasks could check the last emitted state enum
value and either start a new attempt or fail the request.  This approach
also required holding an async lock for the duration of an update
attempt and another to manage the shared state, which was difficult to
reason about.

This change makes a new top-level async task that is responsible for
managing the current update attempt, tracking the current update attempt
state, and forwarding state to monitoring clients.  All requests to
interact with that task now go through a dedicated control channel,
removing the need for the async locks and various generic parameters on
the type used to control the update attempt.

This change also:
* Modifies the update attempt to yield status updates as a stream
  instead of directly emitting state changes to the monitor

* Moves Initiator into fidl_fuchsia_update_ext

* Defines a non-fidl UpdateOptions struct that enforces the type
  invariant that an initiator is required

* Derives Arbitrary on the fidl_fuchsia_update_ext types to simplify
  the proptest helper strategies

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403415
Original-Revision: 3665053352da0f89ec07b34c942bb58d8fee9c49
GitOrigin-RevId: 33e8da483955126c4f7536253661e33bfca15750
Change-Id: Ie8f94e7360268eed216d012f3a303d748da45268