[roll] Roll fuchsia [e2e] Use fake fuchsia.intl.* for chromium tests
This change replaces the test fixture consisting of a real
intl_services.cmx, and its backing services, setui-services and stash
with a fake implementation which has no dependencies.

Currently the fake implementation serves a static initial profile as
specified by startup flags in its manifeste file.

The fake has a bespoke FIDL interface that a test harness can use to
change the locale if needed.  This functionality is currently *not* used
in the tests, but if tests start requiring this, it could be added
fairly easily.  See README.md for more details on how to achieve this:
https://fuchsia.googlesource.com/fuchsia/+/refs/heads/master/garnet/examples/intl/manager/README.md

Added the fake binary into the terminal.gni so that it is available
to the tests that have it as runtime deps.

Original-Bug: 55422,55347,55357
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404080
Original-Revision: 247c12f150515d070380381621c193140d6af45b
GitOrigin-RevId: b31c59c669f061dea4051de587d661dc9740f125
Change-Id: Ia3ffa7f22a7c9273c631c45b5f6ea03a74863665