[roll] Roll fuchsia [bt][hog] Fix HoG and bt-gatt-svc device lifecycle
The bt-gatt-svc device and its children are expected to get removed when
the associated GATT service is removed (following disconnection or a
remote GATT database change). This was not happening for HID services as
the bt-hog driver never removed its devices.

In addition it was possible for GattRemoteServiceDevice::Unbind to be
called twice since the same Unbind() method was called by both the
HostDevice to schedule removal and by the DDK as a side effect of
Unbind() calling device_remove_deprecated() on itself.

* This CL moves bt-gatt-svc lifecycle management entirely into
GattRemoteServiceDevice via DDK and removes the tracking and explicit
destruction from HostDevice.

* GattRemoteServiceDevice now uses ddktl types to reduce DDK
boilerplate.

* hogd_unbind now schedules the removal of both the parent hog device
and the children published for HID boot keyboard and mouse devices.

* hogd_unbind now calls device_unbind_reply as is required

* The device unbind/release hooks for the boot device children are
intentionally left unimplemented as they don't do any work.

Test: Automated tests are currently blocked on test harness support for
      bt-gatt-svc devices (see fxb/2757).
      Manual testing steps:
      1. Pair a BLE keyboard and/or mouse.
      2. Confirm that there is a device published under
      /dev/class/bt-gatt-svc.
      3. Check `dm dump` for a device hierarchy that looks like:
      bt_host > bt_gatt_svc > bt_hog.
      4. Disconnect the keyboard/mouse.
      5. Confirm that no devices are published under
      /dev/class/bt-gatt-svc.
      6. Confirm using `dm dump` that bt_host has no published children.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398575
Original-Revision: d18319c30f1b49d187dfe9e3792ce9d9d9921e57
GitOrigin-RevId: 431602f8789a3dcbca0caacd2e083bb9cfe34017
Change-Id: I61c7ab7d00459f1881dc88330ebdb1cf9275d611