[roll] Roll fuchsia [sshutil] Have NewSFTPClient check if conn is closed
Previously, NewSFTPClient would try to construct a new SFTP client using
the current underlying ssh.Client. But if the connection has been
closed, the connection's Client field will be nil and the method will
result in a segfault. To avoid this, we first explicitly check if the
connection has been closed and return an error if so.

This is unfortunately tricky to test because it would require
implementing the server side of the SFTP connection protocol in
testserver.go.

Original-Bug: 48042
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403876
Original-Revision: 0a8ac6e8b34ee7da2566b924360ce26b211914db
GitOrigin-RevId: d642a1d7b55e319cd212228f4296f2f85b64bd4d
Change-Id: I36e275f1cf9714874b375dd8609dcf734b4ee558