[roll] Roll fuchsia [debugger] Add other thread info for exceptions.
Adds a vector to the exception record to store information for other
threads that were stopped as a result of this exception.

Currently this is not hooked up, but since no other threads are
currently stopped, the values are correct by definition. This will be
hooked up in subsequent patches when we implement the stop mode
controls.

Original-Bug: 55511

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404157
Original-Revision: aa51d1c732ee1335c8c79eab1f9b832ee2e15e33
GitOrigin-RevId: 9e886c7ef11882039c22f426c93edbee939acd42
Change-Id: I8f69ac87fa355314c01b68c10d76ecdd1c499f9a