[roll] Roll external/github.com/catapult-project/catapult dashboard: Be more liberal with bisections
This change reduces the constraints placed on what points/anomalies in
the UI can be bisected. In the past we only allowed bisections on
non-single-change ranges. We're removing that constraint now so that we
can allow single-point bisections for reproducing issues.

In this change we reflect reality that there are no masters and slaves
involved in the implementation of this service. We're removing loaded
terminology and taking a stand that we don't condone and do not support
using terms that have historical baggage.

Let the record reflect that we will do better going forward.

R=abennetts@google.com, fancl@chromium.org

Original-Bug: chromium:1096626
Original-Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2275573
Auto-Submit: Dean Berris <dberris@chromium.org>
Original-Revision: 38d85ba3f9afe5238601f215d97f3b8796998e33
GitOrigin-RevId: b99a19f707c15a9f91f002aa1f2f9968b129adb7
Change-Id: I230dc6762e1a8de4e9baf99c10ad9e782ae4d57d