[roll] Roll fuchsia [cksum] Update for bootloader compatibility
cksum library is needed in bootloader for abr support. Currently kernel libc is
also built with cksum if "is_kernel" flag is set. This flag is set in the x64
bootloader as well and it has it's own libc. This CL resolves this
conflict to enable sharing the library code base for both kernel and
bootloader.

Test: No functionality changes made.
Original-Bug: 55514
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404140
Original-Revision: 6e28ac15b2b5e2ebd57d8e927834e15947387d51
GitOrigin-RevId: 3e7a647d717add1831700f0f942173a409826724
Change-Id: Ie161edac651cf21b6db3f6b66afb053c62d56614