[roll] Roll fuchsia [storage] Fix race between block client terminating and reconnecting.
One of the tests in block-client-tests tears down the connection to the
block server and then immediately reconnects. This fixes a race where
the server hadn't quite processed the fifo closing before trying to
start a new one up, although it should only occur if the client fails
to send FifoClose on the channel. A separate CL follows with a fix for
that.

This isn't easy to add a test case for (as is often the case with
races). To test this fix, I added a sleep call to simulate the server
thread taking its time to terminate and verified this reliably
reproduced the flake without the fix and that the fix worked correctly
with the sleep call.

Original-Fixed: 54733
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403980
Original-Revision: 2f1051f079a197791b088c14ce3efc667eadf8f0
GitOrigin-RevId: 5ca3710d89256660a36ba6de08009eb136ee462c
Change-Id: I9ebc3518c462ad4d87a1e29893da1097f0044bce