[roll] Roll fuchsia [wlan][ralink] Migrate off device_remove_deprecated.
device_unbind_reply is used to reply to the unbind hook,
and device_async_remove is used to trigger a device unbind.

It is guaranteed that the unbind hook will only ever be called
once, and it is required to reply to unbind (otherwise the
device will never be released), so I have removed the early
return from the unbind.

BUG=37831

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402710
Original-Revision: 8bb846dcd815f13b129e31f11f1149af1928c86a
GitOrigin-RevId: 31f34fe410802de6947fe28ac343a515abae5f82
Change-Id: I01bf89c66d5ab6ba9be5890f549979a6852f6d76