[roll] Roll fuchsia [zircon][thread] Restrict thread starting addresses.
This CL disallows setting non-canonical addresses for userspace threads.

The CL adds a `arch_is_valid_user_pc` function that returns whether a
given pc is valid for a userspace program in the current arch.

The 0 address is valid in both x64 and arm64 (used for tests).  Only
userspace addresses are allowed in both x64 and arm64.  In x64 addresses
also have to be canonical.

The check happens in `ThreadDispatcher::MakeRunnable`, which is called
from `ThreadDispatcher::Start`. There's also an assert for that same
check in `arch_enter_uspace`: if we got to that point with an invalid
address, we should just panic.

Without this fix, callers to `zx_thread_start` and `zx_process_stat`
could crash the system in x64 by just setting a non-canonical address.

The problem is that doing an `iret` to a non-canonical address causes an
`#GP` exception, which the kernel can't handle because the `gs` register
is in an invalid state.

To test I used:

```
$ fx set bringup.x64 \
	--with-base //bundles:tools \
	--with-base //garnet/packages/tests:zircon \
	--with-base //bundles/bringup:tests
```

And then:

```
runtests -n core-threads
runtests -n core-process
```

Original-Bug: 54134
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399694
Original-Revision: 5a5b5847f83d08fdb46b0e7fa6cb0209ad697307
GitOrigin-RevId: d4c5b8aa1673b27735dbe6cdadc24c312c78ab3a
Change-Id: Ibb0718fd50f4c48081b4adb11a431d72a59df4ac