[roll] Roll fuchsia [fidlcat] protobuf save/replay
Adds options to save the content of a session to be able to replay it
later.

The session is stored using a protobuf format.

Adds an option to dump the content of a save protobuf file.

Uses a map instead of a vector to store the syscall definitions.

Also creates an event for exceptions.

Displays the column on top of the file name and line for stack frames.

Original-Bug: 54434
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396093
Original-Revision: bc52947d04caf568b6aaa1221c5d48013cb5ec14
GitOrigin-RevId: 227fe9e378a2848b3d9a82b08331575078d7dd19
Change-Id: Ib14dd26b3045ea477a36608f17eb961d31e8b3f0