[roll] Roll fuchsia [lowpan][lowpan-spinel-driver] Reset the send window before calling open.
This change fixes a race condition in `SpinelDeviceSink::open` that
could cause the send window to be reset after the Spinel device had
already (re)opened and sent us a `OnReadyForSendFrames`. This would
cause the send window to remain at zero, resulting in no frames being
able to be sent to the device.

Aside from correcting what is a clear logic error, writing a unit test
for this change that would have failed previously but will now succeed
is tricky. The issue was discovered when testing on real hardware, and
manual verification indicates that this change does address the
problem. Existing tests continue to pass.

Test: fx test lowpan-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401679
Original-Revision: 5ff9ef379311b9edc7a21b641e5044f2309a3113
GitOrigin-RevId: 0792354d2d92a37df22f50538c68781e5f1559f4
Change-Id: Ia07a5ac44bd898909fe3546647cc95983eb85ba4