[roll] Roll fuchsia [bt][hci] Fix a potential overflow
Fix a potential overflow found during a security audit.  See the
referenced bug for details

Currently, there are no good ways to write automated tests for this.
My understanding is that this entire driver is going to be replaced
with a new version sometime soon, so investing a lot of extra effort
in automated testing of this may not be worth it.  I have manually
done a spot check of this patch by deploying to a Sherlock and making
sure that I can still discover, pair with, and communicate to a set of
BT headphones.

Original-Fixed: 55287
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404695
Original-Revision: 5008836c7a75422ba9653f807ede965a3ce7e1b6
GitOrigin-RevId: b6d6bc088dc287f2c36496f3b37943891f640121
Change-Id: I38cead1e75cf2a771c84ad5fba4bb227412c829b