[roll] Roll fuchsia [blobfs] Make next_node undefined at end of list.
The documentation in format.h and the public markdown said that
next_node should be zero in the case of the last node in the linked
list. The header file specified that node zero should be reserved,
making this a safe "null" node, however the documentation didn't include
this detail and we never enforced it anywhere in our code.

This change updates the documentation to match our implementation where
we determine the end of the list based on the extent count in the inode
and derive no meaning from the next_node at the end of a list. We
instead purposely populate next_node there as 0xFFFFFFFF only as an
implementation detail to trigger crashes with high probability when the
pointer is erroneously followed.

Original-Bug: 53860
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402887
Original-Revision: 082d8c5b0885c375206841ae78c1fb631395d4c2
GitOrigin-RevId: adeddef162c76f5b422f25ff1ef962552775bb46
Change-Id: I8b3180f9a96536f148171d4416ac5882c98cfe48