[roll] Roll fuchsia [swd-e2e-tests] Allow omaha server address to be specified
This extends the omaha server to be served on a fixed address, rather
than an arbitrary port. This is helpful for testing omaha, because the
omaha server is embedded into vbmeta. If this port is random, it
requires users to repave the device in order to perform cycle testing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404658
Original-Revision: 64308adedf0790b3a5c154a704e24050374f3aa3
GitOrigin-RevId: ed04043dab1149221125743da18dcf6dab5182e6
Change-Id: I748e70d59dfd360d725772cda8f9d367a467c403