[roll] Roll fuchsia [magma] Fix test flake due to incorrect upper limit.
Fix a flake caused by incorrectly specifying the upper limit as
an absolute address to zx_vmar_allocate, instead of relative to
the base of the parent VMAR.

The upper limit parameter to zx_vmar_allocate remains relative to
to the parent VMAR base, since this is parallel to the semantics of
the original offset parameter when not using the upper limit flag.

Using an absolute uppper limit address is simpler from the Magma
API perspective and is already documented as such. The Zircon
platform impl is updated to perform the translation.

Original-Fixed: 53783
Test: msd_arm_mali_integration_tests.cmx passes.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404919
Original-Revision: 38d65fbd3afca2b666a354150e64f11609df46ba
GitOrigin-RevId: 1ec296b55b1db6ddd2d633719de20bba8ee6c9df
Change-Id: I4b9cc852449d2fce3a622495da358d53bb186406