[roll] Roll fuchsia [a11y] Screen Reader uses the locale ID from the system.
This change turns on the use of the locale ID provided by the system to load the translated strings for that locale.

In order to do that, it packages with the A11YManager all translated strings in the resources section of the package.

Note the subtle change in App tests:
Before, it was using a fake locale because the Screen Reader was not loading strings. Now, it is, so it loads real English strings in this case, using the whole l10n infra.

FIXED: 55181
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404599
Original-Revision: cd338d141f80aa58bfdc4b1b52256008510c815f
GitOrigin-RevId: d788e0794a0a950f65e403fb038fb6e1eb0b0fb5
Change-Id: Id6f33289966a99e8b6702b4245235fb2945a4b09