[roll] Roll fuchsia [load-balancer] migrate fn, handle affine change.
Handle the case, exercised in the set_migrate_fn_test, where the
affinity of a thread changes when a migrate function is set. When
determining if we could stay on the old cpu until delayed migration
happened, in the old path we compared against available_mask in the
new path we also included affinity in that decision which was incorrect.

Also we should be comparing against last_cpu, not initial_cpu. It is not
correct to schedule a migration on a new thread so we add in the bypass
if the last_cpu is unset, which is the new thread case.

This brings us inline with scheduler.cc:650 which is the main flow.

Test: Ran k ut with the load-balancer enabled.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405375
Original-Revision: ec25750cc3043ba31a9c0504eae98c75fca78a20
GitOrigin-RevId: a264fb7a4e2e70b751b0822c6bae98370b8fe6d6
Change-Id: I631b52558d75390547a31baa34754154efffd8a0