[roll] Roll fuchsia [fx][qemu] Tweak qemu arguments to be qemu 2.12 and qemu 5 compatible
A few tweaks:
- switch gic_version argument to gic-version, both of which are accepted
in qemu 2.12 but only the latter accepted in qemu 5
- set the machine emulation from 'virt' to 'virt-2.12' to fix it at 2.12
feature set with regards to memory layout. newer machine emulations
break the current board driver, which will need to be fixed in the
zircon codebase.
- TCG only: default gic-version to max, set cpu to max. Both of which
are supported in qemu 2 and 5. Nets new cpu features in qemu 5, no real
functional change for gic version but is cleaner.

For KVM runs the cpu and gic versions are the same
(gic-version=host,cpu=host) so will not need to be tweaked.

Also removed a hard coded --gic=3 override in the upper
tools/devshell/qemu script that as far as I can tell was unnecessary
anyway because the run-zircon script already defaulted to 3, and now
defaults to max which is effectively 3.

Original-Bug: 54317
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401675
Original-Revision: cdc9836507a6b740d80f65a39ea901ccffc38a60
GitOrigin-RevId: bbb665ec71c1d412d4affaa8019df1f0686c28a3
Change-Id: Ib026d89d96bf026ce40bfc63ba4117453e69622f