[roll] Roll fuchsia [shell] Fix potential race during loop shutdown
The Server class is now provided a loop, instead of creating its
own.

This allows us to use the loop class's support for worker threads to
run a server in a separate thread, instead of the hacky thing we
were doing before.

The bug linked below is a test flake report.  I can't repro, but
it looks as if the loop shutdown races with the destruction of the
loop, causing a crash.  This refactoring makes the shutdown and
the destruction take place in the same thread, which should mitigate
that problem, if that is the problem.

The code is clearly cleaner this way, too.

Original-Bug: 55582
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405536
Original-Revision: 60de82881cf587fc5118ca240afb5b516cff4508
GitOrigin-RevId: 71801a3e9a24c99a76eaf9aef0201649d16de3ba
Change-Id: Ic7f9c7fc4e849254e3e5d035747fc9fa102bd98b