[roll] Roll fuchsia [thinfs] Override NodeGetFlags and NodeSetFlags behavior
The current default behavior of unimplemented transitional FIDL
methods like fuchsia.io.{NodeGetFlags, NodeSetFlags} is to crash the
FIDL server when a client calls them. This is undesirable for
thinfs/pkgfs.

Override these methods (which are the only transitional fuchsia.io
methods) and provide implementations which either map to the GetFlags
implementation, or return an error, but which do not crash thinfs.

Original-Fixed: 50557
Test: Added pkgfs integration tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404907
Original-Revision: b98b7bd4ddc1aefcd75c2dee99230900a85425e3
GitOrigin-RevId: 5d9dda26575300172c94b517167155219bb30466
Change-Id: I05021ecbaf95f8221a18e66f07f3ef06397eff61