[roll] Roll fuchsia [debugger] Add to process handle extraction.
Converts the unwinder to using the ThreadHandle and ProcessHandle abstractions.

Converts thread state to a new class to allow this to be abtracted better.

Removes the process_info file and moves the ELF utilities from there
into a new file. This new file now uses the ProcessHandle abstractions
which will allow the module extration to be tested in the future,
although there are still no tests for this older code.

Creates a new GeneralRegisters container to provide cross-architecture
accesors for the important registers. Use this when relevant and move
the getter/setters of the general registers from the ArchProvider to the
ThreadHandle.

Updates changed function signatures to prefer to return values rather
than use out params, and use references rather than pointers for out
params (as specified by updated style guide).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404359
Original-Revision: b308e02da961f5d7a1eb392524ccf08be9635784
GitOrigin-RevId: 2fcf28cea9a4f6bf1cbcc2444957d439f5cd848d
Change-Id: I73264f3b8aee0118d6511d31405d9698d14d40fe