[roll] Roll fuchsia [debugger] Move debug registers to ThreadHandle.
Remove all variants of GetInfo. The thread variant on the ArchProvider
was already unused. The variant on the ProcessHandle was needed only to
get the return code, so that function was repurposed.

Move the single step and exception conversion to ThreadHandle. This
allows the architecture-specific overrides of ArchProvider to be
removed. Callers are updated to use just ArchProvider. Eventually this
class should be removed since there should be no need for mocking any
archicture-specific conversion functions.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404433
Original-Revision: 00641e6cc9aad6a74e6340bce9d0bfef5860175c
GitOrigin-RevId: 9395e5e25527b0203edbd4a6a16ec077e788b700
Change-Id: Ibfa909dca03b02eb5b7562d8bde45c9693f221ec