[roll] Roll fuchsia [fvm] Name data partitions "minfs"
This corrects a difference between the behavior of the fvm tool when
it's used to create sparse image with an explicit data partition (eg:
//build/images:fvm.sparse.blk) vs with empty minfs (via
--with-empty-minfs, eg: //build/images:fvm.blob.sparse.blk). In the
former case the partition will be named "minfs" whereas in the latter it
would be called "data", resulting in different device paths on the
device.

This is because when supplying a partition image the filesystem type is
sniffed and the name of that filesystem type (minfs or blob) is used as
the partition name wheras in the --with-empty-file case in
SparseContainer::AddCorruptedPartition used kDataTypeName - "data".

Original-Bug: b/159251055
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405477
Original-Revision: d1091e43e88c000980933d619d394dbbd3f606cb
GitOrigin-RevId: 9f0e4f538b6042282331e5e7c21805d683345623
Change-Id: I1da1f8607eccd3c7a47c346e4319d65d8dfcc596