[roll] Roll fuchsia [fidlcat] inference
renames displayed_action inference and calls it when the event is
created.

The inference is now computed on events instead of using a syscall
decoder. One advantage is that we can call it before displaying the
event.

Original-Bug: 54434
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405284
Original-Revision: fb0bfefa029ae1e6fa12b7727d163c289a8bce8c
GitOrigin-RevId: ceb0e247320d74dc26111216cef0ad0cc9b13c0d
Change-Id: Ifd2651aaf0dfec275d4735d9b687200b5e495be9