[roll] Roll fuchsia [lint] use dart from prebuilt
Updates the path for the dart prebuilt to be the one in
prebuilts instead of topaz.

The command was already commented out so was not actually being
used but this is the correct location of the prebuilt when this
is used.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405790
Original-Revision: 56afa66fd01de1839fe34dcbba5cddd490e3a06c
GitOrigin-RevId: 4368e695c0e5342d70a9743ddfaf0e19109edb3a
Change-Id: Ifde629f80306719f2da20f2f1cb33e01276d31a0