[roll] Roll fuchsia [scrutiny] Improve component graph search functionality.
Reimplement how 1 vs many nodes are shown so that uses and offers show
more than just the first matching node. When a single node is found that
matches the search query, show that node and all nodes that are used by
or that use the single node. When multiple nodes are found that match
the search query, only show all of matching nodes.

Since the manipulation of the graph is a little more complicated, added
a search button and only update the graph on <enter> key press or when
the search button is clicked to avoid heavy lag when typing in search
queries in a large graph.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403563
Original-Revision: f6771d8bf1e16e90cdf2fe1fde5e601c29475919
GitOrigin-RevId: ed6dc912dd0dd4637cdc1480819631bbeb71e3dc
Change-Id: If49b764c6d773df970d5cde2893743ebda750494