[roll] Roll fuchsia [fidl_codec] Fix estimating printed size of integers
We estimate the printed size of values to decide if we should print
structs and lists in an expanded form. This makes it important to know
the exact number of columns each value occupies.

Integer values often represent an enum value with a string label. Some
other are printed in hexadecimal format or are padded with zeros.
Therefore the printed representation of an integeral value is often very
different from its raw numerical form.

This patch calculates the accurate size of a value with an underlying
integer encoding.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405975
Original-Revision: b9f1f5c251101e7d0b68a31f25acde75c9909aad
GitOrigin-RevId: 6c25e351917b02825c0354c44b816f8259aaa0a5
Change-Id: If8c5c4f6a5eef429aae474899a003df22b664d7d