[roll] Roll fuchsia [gn][fx][set] Introduce default top-level host-only labels
This extends the build and `fx set` to add support for building labels
that are only built with the host-side toolchain, and are not included as
a package on a target device.

Along the way, this renames the variables `base` and `universe` in
tools/devshell/set in order to make it more clear what these variables
contain.

Test: Added tests to make sure `fx set` will generate the correct args.gni.
Verified that I can specify a host-only label with
`fx set ... --with-host tools/net/sshutil:sshutil_tests` (which happens to
only build on the host-side, since it uses some libraries that haven't been
ported to Fuchsia).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401626
Original-Revision: 417cf8cfb3fc60270b88f879707aed9909523c37
GitOrigin-RevId: b78c003fb2231d40b7e189ece82d2b9213c8995f
Change-Id: I67a7b9cae44a79919da9c0df1a3981dbf206200e