[roll] Roll fuchsia [fbl][containers] Relax ContainableBaseClasses rules
Previously, it was illegal to use ContainableBaseClasses to specify
the containers that your object could exist in if _any_ of those
containers made use of unique_ptr to track the object (unless you
specified the AllowMultiContainerUPtr node option in your object).
This CL relaxes that restriction and allows a user to use CBC with
UPtr provided that there is exactly one container type in the list of
containers.

Tests: updated unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405964
Original-Revision: 5508a75858e5447591bf377bcc7db030ac8190b2
GitOrigin-RevId: f0e82f1192793cf8ecfb53c6c98cbc0a2f00e25d
Change-Id: I51a9abda657401cf307e9c13c283d3ad9a35ba0b