[roll] Roll fuchsia [modular] Workaround for fxbug.dev/55769 in Agent Runner.
This changes the logic for when to use an agent's outgoing dir to
connect to services it publishes versus when to use
fuchsia.modular.Agent to work around a bug in some Directory
implementations.

Verified test hangs without the workaround and passes with it.

Original-Fixed: 55823
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405617
Original-Revision: 0c0844dfed3bf9052201341527e527b0bc9f0d65
GitOrigin-RevId: 147c5be49715cdf364e907f582172ba91f778a69
Change-Id: Id423e1b8ae3df8d35c8faf40354ca5124b540e6d