[roll] Roll fuchsia [fidl][rust] Use unsafe_{encode/decode} for zx_status::Status
These are effectively integers, so the implementation reuses the i32
encoder/decoder. Epitaph is also updated because it only contains a
zx_status.

This saves 12288 bytes in binary size

There aren't benchmarks for this, but when this was done for numbers
elsewhere, it has significant performance improvements. I believe
this is also the case here, because the code size savings is due
to simplified implementation (in integer case, a single instruction).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405634
Original-Revision: db8c1b6d643c5dd062a6cb81295e0ddb97a837df
GitOrigin-RevId: 0b04831da0046d69c1278751cc7421bb38f50367
Change-Id: Ifab67040f816166a116a7ae066ef7fea1e2692c7