[roll] Roll fuchsia [GN SDK] Updating to use common SSH keys
This change updates the path for the SSH key that is used
to connect to Fuchsia devices to be in ~/.ssh/fuchsia_ed25519 and
the authorized_keys file to be ~/.ssh/fuchsia_authorized_keys.

If a shared key does not  exist, but a key does exist in the old location (~/.fuchsia/.ssh/pkey)
it is moved to the new shared location and the authorized_keys file is
regenerated.

If no existing key is found, one is generated.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404914
Original-Revision: 5ac9b6ddcefaa9212cfcf2d236d81b789c04520c
GitOrigin-RevId: 2bc0911e214fd009043e288107c6d6580bb09774
Change-Id: Icf46dd5d041ccace7c758474c965206fa8b4b00c