[roll] Roll fuchsia [settings] Migrate privacy settings back to watch
This a step in the process to rename Watch2 to Watch. A previous api
change required us to add a version of Watch that didn't return a
result/error, so it had to be renamed Watch2 so that both could
exist for a soft-transition. This change is to migrate back to the name
Watch.

Original-Bug: 55685
Test: fx test privacy_settings_ptr_unittest

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405097
Original-Revision: 4a2ea933e2b7c66d37ed5541f09c8c7b9a029069
GitOrigin-RevId: 5bd90fd66919164150fe370933a6a31a036dda7d
Change-Id: I48bc6374eb04496227d5539f25982611228af939