[roll] Roll fuchsia [audio_core] Work around potential race condition
The MixStage will hold a strong reference to any PacketQueues while the
mix is in progress, however the backing Renderer may be free'd in
parallel with the mix job. Currently, PacketQueue returns a reference
clock that is owned by the BaseRenderer, which means this may be freed
after we've released the lock in MixStage.

Work-around this by only reading source->reference_clock() while the
lock is held.

Original-Bug: 55851
Original-Fixed: b/160882220
Test: No behavior change. Verified previous crash that could manifest by
      aggressively scrubbing mute on/off no longer occurs.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405897
Original-Revision: dffffabaf66c78e1d9481b5ea6093955d83a80c5
GitOrigin-RevId: e949bef2a65494cf87ea466dbec6461121c1e30e
Change-Id: I0c27fc81a4a08aefe01cd3b0ac9ce643c6eac3fe