[roll] Roll fuchsia [blktest] Move the blktest library code into the blktest executable
The blktest library is only used by the blktest executable, so it
doesn't need to be separate.  (The blktest executable is invoked by a
couple of different tests, so it is already reusable by that route.)

One reason for doing this is that it simplifies the code.

The immediate motivation for doing this is to allow blktest to be
converted to use zxtest instead of the unittest library.  unittest is
dynamically linked, but currently zxtest is statically linked, which
would cause the blktest executable and library to get their own
separate copies of zxtest, which causes the blktest executable not to
see the tests registered by the blktest library.

Another reason for doing this is that if tests are put in libraries,
there is a risk that the test cases don't get linked in at all.  That
happens if we change the blktest library to be statically linked.

Original-Bug: 51652
Test: blktest is run by ftl-integration-test on bringup builders on CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405382
Original-Revision: a347ae78688b6e55d32879646b1118858adf0b39
GitOrigin-RevId: bdd64922b28904cf1e4c1ed3bbeb640f4d12ab46
Change-Id: Idbd8020b77c1bae711071378ac6251331d3d29b7