[roll] Roll fuchsia [settings] Migrate volume relay back to watch
This a step in the process to rename Watch2 to Watch. A previous api
change required us to add a version of Watch that didn't return a
result/error, so it had to be renamed Watch2 so that both could
exist for a soft-transition. This change is to migrate back to the name
Watch.

Original-Bug: 55685
Test: fx test bt-a2dp-sink-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405076
Original-Revision: f7aed89dc96c5c1dcddc3e1e132e54b868bdd98c
GitOrigin-RevId: f6d967dda81fa7fd3b346207d9e047a633087e6c
Change-Id: I2c6475e2aebcd37861e0b1e9ee3f3177d4559182