[roll] Roll fuchsia [settings] migrate intl back to watch
This a step in the process to rename Watch2 to Watch. A previous api
change required us to add a version of Watch that didn't return a
result/error, so it had to be renamed Watch2 so that both could
exist for a soft-transition. This change is to migrate back to the name
Watch.

Original-Bug: 55685
Test: fx test intl_property_provider_impl_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405096
Original-Revision: 7176b99ddc0da0d0692c0cc36424121f5800f189
GitOrigin-RevId: 25e30da048dd275a6b94de5ca4affc3214d39acb
Change-Id: I2dd8e8161325c0bf6a18c40aa6701a59d107a14e