[roll] Roll fuchsia [build/images] images.json includes API scripts
We do not yet have a defined API for what images are used in what
combination to target a particular device in a particular state. That
roll is fulfilled by pave.sh and friends, which are presently missing
from images.json, resulting in artifactory not uploading them. Users
can not safely consume the images without one of either these scripts
or the aforementioned API that is yet to be produced. This change
introduces those scripts into images.json as an incremental solution.

Original-Fixed: 55755
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405293
Original-Revision: f2b2aff4617d99ab7a770f6270444bc871276269
GitOrigin-RevId: d8d34eb043215d4023912450ed84140bb2d262d8
Change-Id: I2be00b047a1d83a51ac87648413de7994ffbb545