[roll] Roll fuchsia [settings] Migrate user consent back to watch
This a step in the process to rename Watch2 to Watch. A previous api
change required us to add a version of Watch that didn't return a
result/error, so it had to be renamed Watch2 so that both could
exist for a soft-transition. This change is to migrate back to the name
Watch.

Original-Bug: 55685
Test: fx test cobalt_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405099
Original-Revision: 318f1baf7760d164803a2113cb32b10d803136e6
GitOrigin-RevId: d67ef7238eab8d42733458390cf70dd75e370909
Change-Id: I61359dfefab1a6c225bb24c7a880341833186877