[roll] Roll fuchsia [settings] Migrate a11y manager back to watch
This a step in the process to rename Watch2 to Watch. A previous api
change required us to add a version of Watch that didn't return a
result/error, so it had to be renamed Watch2 so that both could
exist for a soft-transition. This change is to migrate back to the name
Watch.

Original-Bug: 55685
Test: fx test a11y_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405075
Original-Revision: cfcd352e0198d407ca2592566a746e2e3b076e70
GitOrigin-RevId: a4269d0306fa7afb75c6dedfed6fdeee247c26e3
Change-Id: I9f028615f334d15c2afef6c87cb6b4e1cc4c1beb