[roll] Roll fuchsia [mdns] throttle address responses
The mdns implementation responds to queries for A and AAAA resource
records, which map the host name to the IPV4 and IPV6 addresses,
respectively, of the host interface. Multicast replies to address
queries should be limited to at most 1/second. This CL adds this
logic to the AddressResponder.

Test: fx test mdns_tests
      unit test added for this specific behavior

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406219
Original-Revision: 0307e8cd051b3160c86edc858b94f29c644ca463
GitOrigin-RevId: a83239fa833f13d7a413dd567405babeff7df596
Change-Id: I9d3b229c385802400f7d86ece95c77b5c4b56231