[roll] Roll fuchsia [as370] Migrate off DdkRemoveDeprecated().
Previously DdkRemoveDeprecated() was used to reply
to the unbind hook and also for the releasing of a device

UnbindTxn.Reply() is now used to reply to the unbind hook.

DdkAsyncRemove() is now used to schedule unbinding and releasing
of a device. Since Dwc2Device does not implement an unbind hook,
this does not change existing behavior.

BUG=37831

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/394738
Original-Revision: 322fafcf4d6556d1fc92c6f155f63ac5f9ab6b6f
GitOrigin-RevId: 8888797a97e64b7f163d8c3940bf47cc73f8ebb2
Change-Id: Ia7af972d7d987ee7608a9a0ad1349850b889bb8a