[roll] Roll fuchsia [host][blobfs] Reduce contention when constructing blob merkle tree
The host blobfs tool reads files, hashes them, and compresses them in a
small collection of threads. Each thread takes a blob to process from a
global queue, processes it, and adds back the hashed and compressed
object to a completed queue.

To reduce contention, have each thread accumulate objects in a local/
per-thread queue and then merge them into the global queue when the
thread is exiting. This avoids a lock acquire-release pair around a
single vector::push_back() and removes a very short unlock-lock cycle
(see the end of the old loop).

As part of fxr/404773, this change and the other parts of that fxr/
reduced runtime of 'blobfs create' by ~2% when packing core.x64
--with-base //bundles/buildbot:core on a 16-vCPU VM on GCE.

Original-Bug: 55297 It would be nice if the Fuchsia build was faster
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405134
Original-Revision: 897bf3434fde2f7013dd5a9e4e2875e7d4c74623
GitOrigin-RevId: 3b66b455eba86893d75b9ac4c1ada59c26b0e3a5
Change-Id: Id32c71348da52a90bf111a9ad8ed3c72aa160d24