[roll] Roll fuchsia [scrutiny] Add command line JSON parsing.
This change enables parsing JSON directly as commands for example:
component.id `{"component_id": 1}`. The upticks must be the first and
last character of the string and can be the only provided argument.
There is no interior escaping in the mvp.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406677
Original-Revision: cdf3b4a86310bd757c6591a34e07853c68395460
GitOrigin-RevId: bfc6cf96a9f95c1079cadbbe2a7f37d15f2969eb
Change-Id: I1ccc7d5977d74f15b3b102bef64ad703a44d6b84