[roll] Roll fuchsia [kernel][vm] Fix argument validation when upper limit flag is set.
Fix slight differences in argument validation of the offset and
size params to vmar_allocate, depending on whether the upper limit
flag is set. The offset and size should be able to reach the end of
a VMAR when the upper limit flag is set.

Test: Added tests pass.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406740
Original-Revision: 1e022dcdf04f7f7a76b68d43de19e157fb9014c4
GitOrigin-RevId: 7f5f6f8975b84b59f132a83646884931982c39e9
Change-Id: Iaa8158917574e388c26867759fb1672f89eb3cfd