[roll] Roll fuchsia [kernel][debuglog] Move the main Debuglog object to LazyInit
This lets us initialize it earlier and as a result safely print earlier
in the kernel. Rearrange the init order in the top of the kernel and
printf to ensure this doesn't break in the future.

Added some constructor trace code, disabled by default, to help with
printing constructor routines as they're called.

Mark FILE object as constexpr to ensure that the default stdout object
is also ready to go from the start. Would be a good candidate for
constinit when the compiler is eventually switched to c++20.

Tested: booted on x86 and arm qemu with and without
kernel.bypass-debuglog=1. Run through CQ to boot on real hardware.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406355
Original-Revision: 4ad1492ac00babfb922b8119dda4db643aa38750
GitOrigin-RevId: 7b1e067f24e85b62c21c948dedf6f069eb87b70f
Change-Id: Idd65a9e01ff1c4e1c65dcefa5d8654216742c95c