[roll] Roll fuchsia [blktest] Convert tests to use zxtest instead of unittest
This was mostly automated, with some manual changes:
 * change helper functions:
    * return void not bool
    * calls use ASSERT_NO_FATAL_FAILURES()
    * use an outparam for the result value for get_testdev()
    * wrap a thread function to achieve the type change
 * move comments about disabled tests
 * use zxtest_run_all_tests() instead of unittest_run_all_tests()

I tried to remove the "#if 0" around the test case
"blkdev_test_multiple" and update its get_testdev() calls.  However,
this test case already failed to compile before this change.  In fact,
it was #if'd out (and wouldn't have compiled if it were enabled) when
it was first added, in this change:
https://fuchsia.googlesource.com/fuchsia/+/4ef35f3b8366d64cccb0fe6e240fb101238d7dfb%5E%21/

Original-Bug: 51652
Test: blktest is run by ftl-integration-test on bringup builders on CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406694
Original-Revision: 9098631257d41a02597cbf5ded25b1c290e205f0
GitOrigin-RevId: 51f6c2c3f4c07419b108a3f95590e4e78b811ab3
Change-Id: I6a0bdbda64842d0d008336c2489959765850e003