[roll] Roll fuchsia [ubench] Add core affinity to ContextSwitch.
The ContextSwitch benchmark appears to be set up to measure the context
switching time between two threads on the same core. However it doesn't
specify to the kernel that the threads should be running on the same
core so it is subject to timing noise depending on where the threads end
up being scheduled.

This change uses affinity to attach each threadpair to a single core to
make the timing more stable and the results more accurate.

Test: ran ubench locally and ran perfcomp on bots.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405083
Original-Revision: 6f8bcbe95bc458f4c19437016e409f207089109b
GitOrigin-RevId: 5eb58351062447d4e19947450b8d761ecdf8ccbd
Change-Id: Ied8a337a5900c6f56f853cd902ed9d31b4b82d33