[roll] Roll fuchsia [settings] Migrate facade back to watch
This a step in the process to rename Watch2 to Watch. A previous api
change required us to add a version of Watch that didn't return a
result/error, so it had to be renamed Watch2 so that both could
exist for a soft-transition. This change is to migrate back to the name
Watch.

Original-Fixed: 55685
Test: fx test sl4f-service-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406540
Original-Revision: 749d7ee1acddae3542fb84db5d2e94b251cd45c6
GitOrigin-RevId: b00dbc0d2cc2da8f15a058bce62f2f5fa940db0b
Change-Id: Ie24e4c1d4059bc8aaed653d182439b6897cca843