[roll] Roll fuchsia [pm] Improve perf of serving list of manifests
In order to support incremental build, serving from a list of manifests
('pm -p') needs improved performance. The existing code takes
about 20 ms to verify each manifest even when there's no change. For a
build with hundreds of packages, this leads to an unacceptable time to
bootstrap a new 'pm serve -p' session.

Loading targets.json takes 99.7% of the time of a noop PublishManifest call.

This CL fixes it by adding a PublishManifests method to repo.go
that takes a list of manifests and loads the targets.json only once at
the beginning and reuses it. For incremental updates, the fsnotify watcher
still loads the targets.json for each package update.

This fix changes per-package publish time from 20ms to ~60us on a common
test scenario of ~1200 packages.

Test: publish_test.go already has integration tests that exercise the code
paths changed in this CL.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403595
Original-Revision: bee9a7ef1fdb3fcff260376fdd61966aedc40dcd
GitOrigin-RevId: 30ce3bf1040167fb8aa24f48d8d7a1aa4192c69b
Change-Id: Ieec73a6082152c2f61930d0e09cff82288c84860