[roll] Roll fuchsia [fx/pkg] deprecate the notion of "push"
The concept of "push" is misleading as users extrapolate that this
means something akin to "install", and make other extrapolations such
as the fileystem underneath a running component could be changed by
this action (which is intentionally not the case).

A new command `fx run` is offered as a replacement for users who want
the one-shot behavior to build and then run a new thing. This command
also introduces a slightly new feature, in that it will attempt to
terminate a previously running instance of the same component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400493
Original-Revision: eb4fa1b7b7b23741a0f7ce57fbdbda8a1913d781
GitOrigin-RevId: 9964ed38741a3288d16ceda2f5e0ef7d3d4a4e1f
Change-Id: I2250b0d532882dcf0a91cf96b461866e64580598