[roll] Roll fuchsia [Security] Create the security codelab.
This CL includes smart-door, smart-door-memory, codelab and solution
components. The smart-door and smart-door-memory is written
intentionally to contain vulnerabilites.

The smart-door is a very simple fake 'smart door' that allows user
to add home member by using username and password, and allows opening
using the username and password. If none of the username matches, it
would check against a hardcoded admin password to see if it is admin
trying to 'open' the door.

It uses 'smart-door-memory' component also in the codelab as the
persistent storage. The 'smart-door-memory' is a simple token
based storage that allows user to read and write to an isolated
storage space referred by the token.

DO NOT COPY ANY OF THE CODE IN THIS COMPONENT!!!

Tested: Unit tests.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383594
Original-Revision: f091c8f31492933a941f6eaf6a43cab6560f9394
GitOrigin-RevId: ebcb489f60079d0f99c75bd5f7ce96875bd69412
Change-Id: I9f9ce15ccf1f921da658a64ff118afabc73880b9