[roll] Roll fuchsia [kernel][autospinlock] Remove assert, add nonnull check
From looking at dissassembly, the assert for non null at the top of
the AutoSpinLock constructor was causing the compiler to generate a
few copies of the function instead of inlining it. Passing a bogus
pointer to is is pretty hard to do and a pretty low risk, so just
replace it with a NONNULL compiler check and enjoy the slightly
better code gen.

No functional change (except removal of an assert).

Tested: built on arm and x86 and validated that the codegen looks nicer.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406357
Original-Revision: c976b71e34aba4c7a6ec65ea0eb58e8b2e115688
GitOrigin-RevId: 8bc31fb017f271424974c307e81183aef0e18351
Change-Id: I42eb28f8b2deafff6d55335bf3597c676b166263