[roll] Roll fuchsia [display] Remove ZX_DEBUG_ASSERT for amlogic RDMA.
This debug assert is unnecessary because the display hardware protocol
is stateless. The only time the assert can fire is when a client is
sending configs without waiting for vsync, and therefore the missed
vsync notifications caused by failing to apply will not cause problems
for the client connection.

The assertion is replaced with Inspect stats that can be used to spot
this issue in the field.

Original-Fixed: 50650
Test: e2e test on a local Astro

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406919
Original-Revision: 7b24c343e065bddd5416e87885536349b077d87e
GitOrigin-RevId: 0a36c4562e95fa95ceb88d51bdfe502b0d2f055f
Change-Id: I8e89b08602a94c686cb22f74c566368820643a3b