[roll] Roll fuchsia [crash_reports] Add Store inspect metrics
This change gets the store to feature parity with the database with
respect to metrics exposed in Inspect. New metrics will be exposed in
the future, but the primary concern at the moment is replacing the
database with the store.

Original-Bug: 47137

TESTED=`fx test crash-reports-tests`

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407140
Original-Revision: 1f088b530799440b59f26d235112cfed505423ce
GitOrigin-RevId: 4c181f3d5844ff4d126fa3fa1368d2bb5f3c1147
Change-Id: Ie977ea75b45c663704accc7a8bc0d532faffd970