[roll] Roll fuchsia [test][futex-ownership] Normalize long timeouts
Go over the ownership code and normalize all of the uses of long
timeouts meant to approximate infinity.

A long time ago, this code used to use a timeout of 1 second for any
operation which should finish eventually.  This timeout value was
supposed to be the test equivalent of a "soft infinity".  In extremely
loaded virtual test environments, however, this can sometimes not be
enough and lead to test flake.

Over the years, some of these flakes have been resolved by increasing
specific timeouts.  Sometimes up to 15 second, other times up to 20,
but at no point were all of the timeouts increased.  This CL goes over
all of the long timeout uses in the test and makes them use all the
same value of 25 seconds (kLongTimeout, defined in utils.h).  In
addition, it modernized the code just a bit by moving away from
zx_time_t and zx_duration_t, preferring the more type-safe C++ verions
(zx::time and zx::duration) instead.

So, why not just wait forever?  Simply put, if the external test
framework watchdog is used to kill the test instead of an internal
timeout, no useful debugging context ever gets captured in the logs,
making it extremely difficult (basically impossible) to debug any
issues which occur on the test bots.  In a perfect world, the test
framework would provide a value for a soft watchdog timeout instead of
having me hardcode one, but right now it does not.  So, for now, 25
seconds it is.

Tests: updated unit tests
Original-Fixed: 54589
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404991
Original-Revision: 5539f18e26ccffcc1d0f935d1b407751bdd70f86
GitOrigin-RevId: 5e744f3357165899393f2a648fe9560a1aa26d72
Change-Id: Ie95f2abe833afc3b78736802dc8712efe58356ec